+++ src/NVP_COM_Common/NVP_Common_List.cs
@@ -2,66 +2,76 @@
 using System;
 using System.Collections.Generic;
 using NVP_Manifest_Creator;
+using System.Collections;
 
 namespace List
 {
     [NVP_Manifest(
-    Text = "Create an empty collection",
-    ViewType = "Default")]
-    [NodeInput("Размер", typeof(int))]
-    public class List_Create : INode
+        Text = "Делает выборку из списка для сравниваемого объекта, если режим = true, то ищется прямое соответствие, если false -- то частичное (только для строк). Если Режим возврата = true, то вернутся только позиции удовлетворяющие запросы, если false -- то не удовлетворяющие",
+        ViewType = "Modifier")]
+    [NodeInput("Исходный список", typeof(object))]
+    [NodeInput("Сравниваемый объект", typeof(object))]
+    [NodeInput("Режим сравнения строк", typeof(bool))]
+    [NodeInput("Режим возврата", typeof(bool))]
+    public class FilterByCondition : INode
     {
-        public List<object> _elements;
         public NodeResult Execute(INVPData context, List<NodeResult> inputs)
         {
-            int list_size = (int)inputs[0].Value;
-            _elements = new List<object>(list_size);
-            return new NodeResult(_elements); ;
+            List<object> items = inputs[0].Value as List<object>;
+            List<bool> results = new List<bool>();
+
+            object comparing_object = inputs[1].Value;
+            bool comparing_mode = (bool)inputs[2].Value;
+            bool return_data = (bool)inputs[3].Value;
+
+            for (int i = 0; i < items.Count; i++)
+            {
+                var item = items[i];
+                bool result = false;
+                if (item != null)
+                {
+                    //bool can_add = false;
+                    if (item.GetType().Equals(typeof(string)))
+                    {
+                        if (comparing_mode && item.ToString() == comparing_object.ToString()) result = true;
+                        else if (!comparing_mode && item.ToString().Contains(comparing_object.ToString())) result = true;
+                    }
+                    else if (item.Equals(comparing_object)) result = true;
+
+                    
+                }
+
+                if (!return_data) result = !result;
+                results.Add(result);
+            }
+
+            return new NodeResult(results); ;
         }
     }
 
     [NVP_Manifest(
-    Text = "Add element in collection and return the edited original collection",
-    ViewType = "Data")]
-    [NodeInput("List", typeof(object))]
-    [NodeInput("Any object", typeof(object))]
-    public class List_Add : INode
+            Text = "Делает выборку из списка по заданной маске (одноразмерному списку bool)",
+            ViewType = "Modifier")]
+    [NodeInput("Исходный список", typeof(object))]
+    [NodeInput("Маска", typeof(List<bool>))]
+    public class FilterByBoolMask : INode 
     {
         public NodeResult Execute(INVPData context, List<NodeResult> inputs)
         {
-            List<object> elements = inputs[0].Value as List<object>;
-            elements.Add(inputs[1].Value);
-            return new NodeResult(elements);
+            IList items = (IList)inputs[0].Value;
+            List<bool> mask = inputs[1].Value as List<bool>;
+
+            List<object> results = new List<object>();
+
+            for (int i = 0; i < items.Count; i++)
+            {
+                if (mask[i]) results.Add(items[i]);
+            }
+
+            return new NodeResult(results); ;
         }
     }
 
-    [NVP_Manifest(
-    Text = "Get all elements",
-    ViewType = "Data")]
-    [NodeInput("List", typeof(object))]
-    public class List_GetElements : INode
-    {
-        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
-        {
-            //return new NodeResult(((dynamic)inputs[0].Value)._elements);
-            List<object> elements = inputs[0].Value as List<object>;
-            return new NodeResult(elements);
-        }
-    }
 
-    [NVP_Manifest(
-    Text = "Convert List to Array",
-    ViewType = "Modifier")]
-    [NodeInput("List", typeof(object))]
-    public class List_ToArray : INode
-    {
-        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
-        {
-            List<object> elements = inputs[0].Value as List<object>;
-            //return new NodeResult(((dynamic)inputs[0].Value)._elements.ToArray());
-            //List<object> elements = inputs[0].Value as List<object>;
-            return new NodeResult(elements.ToArray());
-        }
-    }
 
 }