+++ src/NVP_nanoCAD_COM/OdaX/AcadBlock.cs
@@ -1821,4 +1821,39 @@ namespace OdaX.AcadBlock
 
         }
     }
+
+    [NVP_Manifest(
+       Text = "Возвращает все Объекты модели ModelStudioCS в данном Блоке в виде OdaX.AcadEntity",
+       ViewType = "Data")]
+    [NodeInput("AcadBlock", typeof(object))]
+    public class GetAll_EntitiesMST : INode
+    {
+        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
+        {
+            var _input0 = ((dynamic)inputs[0].Value)._i as OdaX.IAcadBlock;
+            List<OdaX.AcadEntity.AcadEntity_Constructor> items = new List<OdaX.AcadEntity.AcadEntity_Constructor>();
+            for (int item_counter = 0; item_counter < _input0.Count; item_counter++)
+            {
+                var item = new BaseClass();
+                item._i = _input0.Item(item_counter);
+                bool can_add = false;
+                try
+                {
+                    var mst_elem = item._i.Element;
+                    if (mst_elem != null) can_add = true;
+
+                }
+                catch { }
+
+                if (can_add)
+                {
+                    OdaX.AcadEntity.AcadEntity_Constructor elem = new OdaX.AcadEntity.AcadEntity_Constructor();
+                    elem._i = item._i;
+                    items.Add(elem);
+                }
+
+            }
+            return new NodeResult(items);
+        }
+    }
 }